extract components from Modal
